revert: "feat(conversation): :sparkles: replyWithDirectives on conversation"
This reverts commit f37c51ee6443f7e6e3f681bcd953a023603db971.
