+++ src/conversation/OutgoingConversation.ts
@@ -15,8 +15,31 @@ export default class OutgoingConversation extends Conversation {
     super(eris);
     this.target = target;
 
+    // Join directives with \n\n, splitting every 1900 characters, and adding an `Also` directive if needed
+    const messages: string[] = [];
+    let currentMessage = this.target ? `<@!${this.target}> ` : '';
+    for (const directive of directives) {
+      if (currentMessage.length + directive.length > 1900) {
+        messages.push(`${currentMessage}\n\nAlso`);
+        currentMessage = '';
+      }
+      currentMessage += `${directive}\n\n`;
+    }
+    messages.push(currentMessage);
+
+    // Send each message one by one, all replying to the first one
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    this.replyWithDirectives(directives, channel);
+    (async () => {
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      const reference = await channel.send(messages.shift()!);
+      this.eris.conversationManager.conversations[reference.id] = this;
+      (reference as ErisMessage).eris = this.eris;
+      this.reference = reference as ErisMessage;
+
+      for (const message of messages) {
+        await this.reference.reply(message);
+      }
+    })();
   }
 
   isWaitingForReply() {
